multiBalance to shared
