+++ btc/handlers/btc.handlers.go
@@ -7,7 +7,7 @@ import (
 
 	"sync"
 
-	"github.com/button-tech/utils-node-tool/btc/handlers/multiBalance"
+	"github.com/button-tech/utils-node-tool/shared/multiBalance"
 	"github.com/button-tech/utils-node-tool/shared/responseModels"
 	"github.com/button-tech/utils-node-tool/shared/db"
 	"github.com/gin-gonic/gin"
@@ -175,7 +175,7 @@ func GetBalances(c *gin.Context) {
 
 	for i := 0; i < len(request.AddressesArray); i++ {
 		wg.Add(1)
-		go multiBalance.Worker(&wg, request.AddressesArray[i], balances)
+		go multiBalance.BtcWorker(&wg, request.AddressesArray[i], balances)
 	}
 	wg.Wait()
