+++ ltc/handlers/ltc.handlers.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/button-tech/utils-node-tool/shared/db"
-	"github.com/button-tech/utils-node-tool/ltc/handlers/multi-balance"
+	"github.com/button-tech/utils-node-tool/shared/multiBalance"
 	"github.com/button-tech/utils-node-tool/shared/responseModels"
 	"github.com/gin-gonic/gin"
 	"github.com/imroc/req"
@@ -128,7 +128,7 @@ func GetBalances(c *gin.Context) {
 
 	for i := 0; i < len(request.AddressesArray); i++ {
 		wg.Add(1)
-		go multiBalance.Worker(&wg, request.AddressesArray[i], balances)
+		go multiBalance.LtcWorker(&wg, request.AddressesArray[i], balances)
 	}
 	wg.Wait()
