--- waves/handlers/storage/storage.go
@@ -1,9 +1,0 @@
-package storage
-
-import "os"
-
-var WavesURL = os.Getenv("WAVES_NODE")
-
-type BalanceData struct {
-	Balance int64 `json:"balance"`
-}
